Update README.md
Added information on resetting the tomogram's origin stored in the header. Old tomograms had origins zeroed out, the new do not. Only ChimeraX apparently cares about this, so it could cause confusions when visualizing the annotations. Also changed the steps to open the tomograms as in the latest version of ArtiaX, opening of the tomograms fails.